+++ drizzle/meta/_journal.json
@@ -22,6 +22,13 @@
       "when": 1753434786323,
       "tag": "0002_nifty_hitman",
       "breakpoints": true
+    },
+    {
+      "idx": 3,
+      "version": "7",
+      "when": 1753686690464,
+      "tag": "0003_whole_lionheart",
+      "breakpoints": true
     }
   ]
 }
