+++ src/app/(auth)/onboarding/avatar-upload.tsx
