+++ src/app/(auth)/sign-in/page.tsx
@@ -2,7 +2,7 @@ import { SignInDialog } from "./dialog";
 
 export default function SignInPage() {
   return (
-    <main className="-mt-32 flex items-center justify-center md:h-screen">
+    <main className="-mt-32 flex w-full items-center justify-center md:h-screen">
       <SignInDialog />
     </main>
   );
