+++ src/app/(auth)/sign-up/page.tsx
@@ -1,7 +1,7 @@
 import { SignUpDialog } from "./dialog";
 export default async function SignUpPage() {
   return (
-    <section className="md:-mt-32 flex items-center justify-center md:h-screen">
+    <section className="md:-mt-32 flex w-full items-center justify-center md:h-screen">
       <SignUpDialog />
     </section>
   );
