+++ src/app/layout.tsx
@@ -2,10 +2,8 @@ import "~/styles/globals.css";
 
 import type { Metadata } from "next";
 import { Geist } from "next/font/google";
-import { Toaster } from "~/components/ui/sonner";
-import Header from "./_components/header";
-import Sidebar from "./_components/sidebar";
 import { Providers } from "./providers";
+import { Toaster } from "~/components/ui/sonner";
 
 export const metadata: Metadata = {
   title: "Da Box - Cloud Storage",
@@ -25,18 +23,10 @@ export default function RootLayout({
     <html lang="en" className={`${geist.variable}`}>
       <body>
         <Providers>
-          <div className="min-h-screen ">
-            <div className="flex bg-background text-foreground">
-              <Sidebar />
-
-              <main className="ml-64 flex-1">
-                <Header />
-
-                <div className="p-6">{children}</div>
-              </main>
-              <Toaster />
-            </div>
+          <div className="flex min-h-screen w-full bg-background text-foreground">
+            {children}
           </div>
+          <Toaster />
         </Providers>
       </body>
     </html>
